Refactor variant selection code
