+++ src/server/api/epicRouter.ts
@@ -13,9 +13,7 @@ import type {
     WeeklyArticleLog,
 } from '../../shared/types';
 import { hideSRMessagingForInfoPageIds } from '../../shared/types';
-import type { BanditData } from '../bandit/banditData';
 import type { ChannelSwitches } from '../channelSwitches';
-import { selectAmountsTestVariant } from '../lib/ab';
 import { getDeviceType } from '../lib/deviceType';
 import { baseUrl } from '../lib/env';
 import type { TickerDataProvider } from '../lib/fetchTickerData';
@@ -25,6 +23,8 @@ import { getQueryParams } from '../lib/params';
 import type { SuperModeArticle } from '../lib/superMode';
 import { buildEpicCampaignCode } from '../lib/tracking';
 import type { Debug } from '../tests/epics/epicSelection';
+import { selectAmountsTestVariant } from '../selection/ab';
+import type { BanditData } from '../selection/banditData';
 import { findForcedTestAndVariant, findTestAndVariant } from '../tests/epics/epicSelection';
 import { logWarn } from '../utils/logging';
 import type { ValueProvider } from '../utils/valueReloader';
