+++ src/server/tests/gutters/gutterSelection.ts
@@ -6,9 +6,9 @@ import type {
     GutterVariant,
     UserDeviceType,
 } from '../../../shared/types';
-import { selectVariantUsingMVT } from '../../lib/ab';
 import type { TestVariant } from '../../lib/params';
 import { audienceMatches, correctSignedInStatus, pageContextMatches } from '../../lib/targeting';
+import { selectVariantUsingMVT } from '../../selection/ab';
 
 const moduleName = 'Gutter';
